backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
chromadb==1.0.15
sentence-transformers==5.0.0
fastembed==0.7.1
typing-extensions>=4.12.0
python-dotenv==1.0.1
orjson==3.10.5
sqlalchemy==2.0.41 
asyncpg==0.30.0 
pgvector==0.4.1
pydantic==2.11.7
pytest==8.4.1
pytest-asyncio==1.1.0
pydantic-settings==2.10.1
